split domain workflows and test activity recommendations
